+++ Revo.Infrastructure/DataAccess/Migrations/DatabaseMigrationsConfiguration.cs
@@ -1,17 +1,16 @@
 ﻿using System.Collections.Generic;
-using MoreLinq.Extensions;
 using Revo.Infrastructure.DataAccess.Migrations.Providers;
 
 namespace Revo.Infrastructure.DataAccess.Migrations
 {
     public class DatabaseMigrationsConfiguration : IDatabaseMigrationsConfiguration
     {
-        public HashSet<DatabaseMigrationSpecifier> MigrateOnlySpecifiedModules { get; set; }
+        public List<DatabaseMigrationSearchSpecifier> MigrateOnlySpecifiedModules { get; set; }
         public bool? ApplyMigrationsUponStartup { get; set; }
         public string[] EnvironmentTags { get; set; } = new string[0];
         public IDatabaseMigrationScripter OverrideDatabaseMigrationScripter { get; set; }
 
-        IReadOnlyCollection<DatabaseMigrationSpecifier> IDatabaseMigrationExecutionOptions.
+        IReadOnlyCollection<DatabaseMigrationSearchSpecifier> IDatabaseMigrationExecutionOptions.
             MigrateOnlySpecifiedModules => MigrateOnlySpecifiedModules;
     }
 }