+++ Tests/Revo.Infrastructure.Tests/DataAccess/Migrations/DatabaseMigrationExecutorTests.cs
@@ -57,13 +57,63 @@ namespace Revo.Infrastructure.Tests.DataAccess.Migrations
             var migrations = new[] {Substitute.For<IDatabaseMigration>()};
 
             migrationRegistry.GetAvailableModules().Returns(new[] {"appModule1"});
-            migrationExecutionOptions.MigrateOnlySpecifiedModules.Returns((IReadOnlyCollection<DatabaseMigrationSpecifier>) null);
+            migrationExecutionOptions.MigrateOnlySpecifiedModules.Returns((IReadOnlyCollection<DatabaseMigrationSearchSpecifier>) null);
 
             migrationSelector.SelectMigrationsAsync(
                 Arg.Is<DatabaseMigrationSpecifier[]>(x =>
                     x.SequenceEqual(new[] {new DatabaseMigrationSpecifier("appModule1", null)})),
                 Arg.Is<string[]>(x => x.Length == 2 && x.Contains("providerTag") && x.Contains("optionsTag")))
                 .Returns(migrations.Select(x => new SelectedModuleMigrations(new DatabaseMigrationSpecifier("appModule1", null), migrations)).ToArray());
+
+            migrationProviders[0].SupportsMigration(migrations[0]).Returns(true);
+
+            await sut.ExecuteAsync();
+
+            migrationProviders[0].Received(1).ApplyMigrationsAsync(migrations);
+        }
+
+        [Fact]
+        public async Task ExecuteAsync_MatchesModulesByWildcard()
+        {
+            var migrations = new[] { Substitute.For<IDatabaseMigration>() };
+
+            migrationRegistry.SearchModules("app-*").Returns(new[] { "appmodule-foo" });
+            migrationExecutionOptions.MigrateOnlySpecifiedModules.Returns(new[]
+            {
+                new DatabaseMigrationSearchSpecifier("app-*", null)
+            });
+
+            migrationSelector.SelectMigrationsAsync(
+                    Arg.Is<DatabaseMigrationSpecifier[]>(x =>
+                        x.SequenceEqual(new[] { new DatabaseMigrationSpecifier("appmodule-foo", null) })),
+                    Arg.Any<string[]>())
+                .Returns(migrations.Select(x => new SelectedModuleMigrations(new DatabaseMigrationSpecifier("appmodule-foo", null), migrations)).ToArray());
+
+            migrationProviders[0].SupportsMigration(migrations[0]).Returns(true);
+
+            await sut.ExecuteAsync();
+
+            migrationProviders[0].Received(1).ApplyMigrationsAsync(migrations);
+        }
+
+        [Fact]
+        public async Task ExecuteAsync_MatchesModulesByWildcard_NoDuplicateMigrations()
+        {
+            var migrations = new[] { Substitute.For<IDatabaseMigration>() };
+
+            migrationRegistry.SearchModules("app-*").Returns(new[] { "appmodule-foo" });
+            migrationRegistry.SearchModules("*").Returns(new[] { "appmodule-foo" });
+            migrationExecutionOptions.MigrateOnlySpecifiedModules.Returns(new[]
+            {
+                new DatabaseMigrationSearchSpecifier("app-*", null),
+                new DatabaseMigrationSearchSpecifier("*", null)
+            });
+
+            migrationSelector.SelectMigrationsAsync(
+                    Arg.Is<DatabaseMigrationSpecifier[]>(x =>
+                        x.SequenceEqual(new[] { new DatabaseMigrationSpecifier("appmodule-foo", null) })),
+                    Arg.Any<string[]>())
+                .Returns(migrations.Select(x => new SelectedModuleMigrations(new DatabaseMigrationSpecifier("appmodule-foo", null), migrations)).ToArray());
 
             migrationProviders[0].SupportsMigration(migrations[0]).Returns(true);